Add lifecycle class
